Development

Approved-by: David Gonzalez <david@tailrecursive.co>
Approved-by: Eduardo Quintero <eduardo@tailrecursive.co>

+++ src/components/DesignCenterCustomize/Render3D/index.js
@@ -38,7 +38,8 @@ import {
   TurnOffHintRow,
   MobileContainer,
   MobileButton,
-  MobileButtonWrapper
+  MobileButtonWrapper,
+  MobileHintIcon
 } from './styledComponents'
 import {
   viewPositions,
@@ -110,6 +111,7 @@ import rightIcon from '../../../assets/Cube_right.svg'
 import backIcon from '../../../assets/Cube_back.svg'
 import topIcon from '../../../assets/Cube-Top.svg'
 import hintImg from '../../../assets/designCenterhelpHint.jpg'
+import mobileHintImg from '../../../assets/designCenterhelpMobileHint.png'
 import helpTooltip from '../../../assets/tooltip.svg'
 
 const cubeViews = [backIcon, rightIcon, frontIcon, leftIcon, topIcon]
@@ -1199,6 +1201,21 @@ class Render3D extends PureComponent {
               <FormattedMessage {...messages.drag} />
             </DragText>
           )}
+          <HelpModal
+            open={showHelpModal}
+            withLogo={false}
+            requestClose={this.handleHelpModal}
+          >
+            <HintModalImage src={mobileHintImg} alt="" />
+            {!showHint && (
+              <TurnOffHintRow>
+                <Checkbox onChange={this.disableHelpModal}>
+                  {formatMessage(messages.turOffHintMobile)}
+                </Checkbox>
+              </TurnOffHintRow>
+            )}
+          </HelpModal>
+          <MobileHintIcon src={helpTooltip} onClick={this.handleHelpModal} />
         </MobileContainer>
       )
     }
@@ -1469,7 +1486,8 @@ class Render3D extends PureComponent {
           fileId,
           src: fileUrl,
           scaleX,
-          scaleY
+          scaleY,
+          lock: false
         }
         position.scaleX = scaleX
         position.scaleY = scaleY
@@ -1511,6 +1529,7 @@ class Render3D extends PureComponent {
     } else {
       const { onApplyCanvasEl } = this.props
       const id = idElement || shortid.generate()
+
       txtEl = new fabric.Text(text, {
         id,
         hasRotatingPoint: false,
@@ -1533,12 +1552,17 @@ class Render3D extends PureComponent {
         this.canvasTexture.setActiveObject(txtEl)
       }
       this.canvasTexture.renderAll()
+      let activeElementId
+      if (activeEl && activeEl.type === CanvasElements.Text) {
+        activeElementId = activeEl.id
+      }
 
       if (!idElement) {
         const el = {
-          id: activeEl ? activeEl.id : txtEl.id,
+          id: activeElementId || txtEl.id,
           text,
-          textFormat: style
+          textFormat: style,
+          lock: false
         }
         onApplyCanvasEl(el, CanvasElements.Text, !!activeEl, {
           src: text,
@@ -1613,7 +1637,8 @@ class Render3D extends PureComponent {
           fill: BLACK,
           stroke: BLACK,
           strokeWidth: 0,
-          ...style
+          ...style,
+          lock: false
         }
         position.scaleX = scaleX
         position.scaleY = scaleY
@@ -1706,7 +1731,9 @@ class Render3D extends PureComponent {
             stroke,
             strokeWidth,
             textAlign,
-            charSpacing
+            charSpacing,
+            fontSize,
+            lineHeight
           } = el
           canvasObject.src = text
           canvasObject.style = {
@@ -1715,7 +1742,9 @@ class Render3D extends PureComponent {
             stroke,
             strokeWidth,
             textAlign,
-            charSpacing
+            charSpacing,
+            fontSize,
+            lineHeight
           }
         }
         break
@@ -1898,7 +1927,9 @@ class Render3D extends PureComponent {
       return
     }
     const {
-      responsive: { tablet }
+      responsive: { tablet },
+      selectedElement,
+      canvas
     } = this.props
     let clientX = evt.clientX
     let clientY = evt.clientY
@@ -1961,8 +1992,16 @@ class Render3D extends PureComponent {
         })
       } else {
         if (activeEl && !this.dragComponent) {
+          const selectedGraphicElement =
+            canvas.image[selectedElement] ||
+            canvas.path[selectedElement] ||
+            canvas.text[selectedElement]
           const action = clickOnCorner(activeEl.oCoords, uv)
-          if (action) {
+          if (
+            action &&
+            selectedGraphicElement &&
+            !selectedGraphicElement.lock
+          ) {
             switch (action) {
               case DELETE_ACTION:
                 this.deleteElement(activeEl)
@@ -2077,7 +2116,9 @@ class Render3D extends PureComponent {
   onMouseMove = evt => {
     evt.preventDefault()
     const {
-      responsive: { tablet }
+      responsive: { tablet },
+      canvas,
+      selectedElement
     } = this.props
     let clientX = evt.clientX
     let clientY = evt.clientY
@@ -2114,32 +2155,38 @@ class Render3D extends PureComponent {
           currentTransform
         } = this.dragComponent
         const uv = intersects[0].uv
-        switch (action) {
-          case DRAG_ACTION: {
-            const left = uv.x * CANVAS_SIZE - differenceX
-            const top = (1 - uv.y) * CANVAS_SIZE - differenceY
-            activeEl.set({ left, top }).setCoords()
-            this.canvasTexture.renderAll()
-            break
+        const selectedGraphicElement =
+          canvas.image[selectedElement] ||
+          canvas.path[selectedElement] ||
+          canvas.text[selectedElement]
+        if (selectedGraphicElement && !selectedGraphicElement.lock) {
+          switch (action) {
+            case DRAG_ACTION: {
+              const left = uv.x * CANVAS_SIZE - differenceX
+              const top = (1 - uv.y) * CANVAS_SIZE - differenceY
+              activeEl.set({ left, top }).setCoords()
+              this.canvasTexture.renderAll()
+              break
+            }
+            case SCALE_ACTION: {
+              const cX = uv.x * CANVAS_SIZE
+              const cY = (1 - uv.y) * CANVAS_SIZE
+              this.scaleObject(cX, cY, currentTransform)
+              this.forceUpdate()
+              this.canvasTexture.renderAll()
+              break
+            }
+            case ROTATE_ACTION: {
+              const cX = uv.x * CANVAS_SIZE
+              const cY = (1 - uv.y) * CANVAS_SIZE
+              this.rotateObject(cX, cY, currentTransform)
+              this.forceUpdate()
+              this.canvasTexture.renderAll()
+              break
+            }
+            default:
+              break
           }
-          case SCALE_ACTION: {
-            const cX = uv.x * CANVAS_SIZE
-            const cY = (1 - uv.y) * CANVAS_SIZE
-            this.scaleObject(cX, cY, currentTransform)
-            this.forceUpdate()
-            this.canvasTexture.renderAll()
-            break
-          }
-          case ROTATE_ACTION: {
-            const cX = uv.x * CANVAS_SIZE
-            const cY = (1 - uv.y) * CANVAS_SIZE
-            this.rotateObject(cX, cY, currentTransform)
-            this.forceUpdate()
-            this.canvasTexture.renderAll()
-            break
-          }
-          default:
-            break
         }
       }
     } else if (!!intersects.length && !!this.state.canvasEl) {
